ref: component, template 분리
