+++ src/core/templates.mjs
@@ -1,4 +1,4 @@
-export function drawCars(names) {
+export function carListTemplate(names) {
   return `<div class="mt-4 d-flex">
             ${names
               .map(
@@ -9,27 +9,5 @@ export function drawCars(names) {
         </div>`;
 }
 
-export const forwardIcon = `<div class="forward-icon mt-2">⬇️️</div>`;
+export const forwardIconTemplate = `<div class="forward-icon mt-2">⬇️️</div>`;
 
-/**
- * @param {string[]} winners
- * @param {{onResetButtonClick: Function}} options
- */
-export function TemplateRaceResult(winners, { onResetButtonClick }) {
-    const fragment = document.createDocumentFragment();
-    const $raceResultContainer = document.createElement('div');
-
-    $raceResultContainer.innerHTML = `<div>
-      <h2>🏆 최종 우승자: <span class="winners">${winners.join(', ')}</span> 🏆</h2>
-      <div class="d-flex justify-center">
-        <button type="button" class="btn btn-cyan">다시 시작하기</button>
-      </div>
-    </div>`;
-    $raceResultContainer.addEventListener('click', (e) => {
-        if (e.target.matches('button')) {
-            onResetButtonClick();
-        }
-    });
-    fragment.appendChild($raceResultContainer);
-    return fragment;
-}
